requirements.txt
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2
Django>=4.0
djangorestframework
psycopg2-binary
django-cors-headers

